FIX: BaseBlob.words(self) incorrect docstring
The docstring for `BaseBlob.words(self)` was incorrectly claiming to return a list of tokens **without** punctuation characters and advised users to use `BaseBlob.tokens(self)` if punctuation was required. In reality, `BaseBlob.words(self)` is simply an alias for `BaseBlob.tokens(self)`.

In the name of backwards compability (considering that there are no bug reports of this in the GitHub Issues) this commit simply corrects the documentation rather than changing the implementation.